requirements.txt
numpy
pandas
matplotlib
paho-mqtt
scikit-learn
streamlit
flask
plotly
influxdb-client
dash
pyqtgraph
